add tests to create a user
