+++ package.json
@@ -3,14 +3,15 @@
   "main": "server.js",
   "scripts": {
     "pretest": "npm install",
+    "init" : "node bin/db.js",
     "test": "istanbul cover --print details _mocha -- -R spec tests/** --recursive",
     "start": "node server.js"
   },
   "dependencies": {
-    "express": "*",
     "body-parser": "*",
-    "sequelize": "*",
+    "express": "*",
     "mysql": "*",
+    "sequelize": "*",
     "sqlite3": "*"
   },
   "devDependencies": {
@@ -18,6 +19,7 @@
     "istanbul": "0.4.3",
     "mocha": "2.4.5",
     "nock": "8.0.0",
-    "sinon": "1.17.3"
+    "sinon": "1.17.3",
+    "supertest": "1.2.0"
   }
 }
